# Conflicts:
#   backend/src/main/java/com/dopamines/backend/account/repository/AccountRepository.java
#   backend/src/main/java/com/dopamines/backend/account/security/CustomAuthorizationFilter.java

+++ backend/src/main/java/com/dopamines/backend/plan/repository/PlanRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface PlanRepository extends JpaRepository<Plan, Integer> {
-    Optional<Plan> findById(Integer planId);
+public interface PlanRepository extends JpaRepository<Plan, Long> {
+    Optional<Plan> findById(Long planId);
 
 }
